move lark transformer guts from __init__ to _parser
